com.unity.nuget.newtonsoft-json@1.0.0-preview.2 # Changelog All notable changes to this package will be documented in this file.
The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).

## [1.0.0-preview.2] - 2019-06-21

### This is the second preview release of *Unity Package Nuget.Newtonsoft.Json*.

The name of the package has been changed to avoid a naming issue with npm.

## [1.0.0-preview.1] - 2019-06-21

### This is the first preview release of *Unity Package Nuget.Newtonsoft.Json*.

This is the first preview of a the custom Unity package for Newtonsoft.Json. Please report any bugs.